Use ConfigureAwait(false) for all of the library code
Although we can be relatively sure about the environment, we shouldn't
make any assumptions. Therefore, we enforce the usage of
ConfigureAwait(false) for the code in Anexia.E5E.

Why? The thread context is not needed for our usage. Everything of our
background service (namely E5ECommunicationService) does not care about
the actual context. We neither have a HttpContext that we need to care
about nor something else that's thread specific - usually. However,
there's already one case where we are affected by the usage of
ConfigureAwait and that's our XUnit tests. The previous code (without
ConfigureAwait) caused a deadlock if the test concurrency is limited,
either by hardware (CPU constraints) or by the configuration
(xunit.runner.json). The usage of ConfigureAwait(false) circumvents
this issue.

Further reading:
- https://devblogs.microsoft.com/dotnet/configureawait-faq/