Increase predictability of tests
The old tests were working fine on our local machine, but not on GitHub
Actions. The most notable problem was that the asynchronous code was
working in a different order on GH than locally. This is due to the fact
that the GitHub Actions Runner had a lower CPU count than the local
machine. A quick test with disabled parallelism showed the same
behaviour locally.

As a result, all integration tests were rewritten to make use of the
IAsyncLifetime and class fixtures. However, that alone still could've
caused issues with the tests running into a deadlock. In order to debug
them better, a custom logger was created that prefixes each line with
the current test.

Furthermore, several monitorings are added during startup of the test
host in order to avoid them running for an infinite amount of time.