Move endpoint initialization into build stage
The previous way was not intuitive and caused even myself as the author
to write a bug in the starter kit. If you've wanted to use dependency
injection (DI) before, you had to register your endpoints twice: once
in the build stage of the host, and a second time after that.

This is now refactored into a streamlined IE5EEntrypointBuilder which
allows a better and less error-prone way for registering entrypoints.

Since this was a rather big API change, I also cleaned up the tests in
this commit.