SIANXSVC-1203: Ensure library_version has correct value (#8)
Several things had to get fixed in order for the library version to
appear in the metadata. First and foremost, the GitHub Actions workflow
was setting the PackageVersion (which is used for the NuGet package),
but not the AssemblyVersion. This is now fixed by passing the tag via
/p:Version.

Second, the current retrieved version included the commit hash, which
would have ended up to conflict with the database constraints of E5E
which limit the version to 20 characters. In order to circumvent this
problem, the hash is trimmed from the version.

Because I had the time to work on this part of the code, I took the
opportunity and made the E5ERuntimeMetadata sealed. Technically, this is
a breaking change, however I don't expect that anyone is affected by it,
so it's fine to put it in a patch release.

Closes SIANXSVC-1203