SIANXSVC-1214: Use lcov for CodeClimate reports
Even though the Code Coverage report is correct and other platforms like
GitLab do not have a problem understanding it, the CodeClimate reporter
struggles with double slashes that seem to be introduced by itself.

This is also tracked in an issue, which is open since three years with
no intent of fixing it. Therefore, I assume this is not going to be
closed soon.

https://github.com/codeclimate/test-reporter/issues/458

As a workaround, we're now using the lcov format.